# [PERF] stock: optimize method `_unlink_processed_orderpoints`