@robodoo 2025-03-07
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/200852.png)](https://mergebot.odoo.com/odoo/odoo/pull/200852)