@adda-odoo 2025-03-11
@Aurelienvd retargeted to 16.0 here --> https://github.com/odoo/odoo/pull/201243